Move extern specs tests
