+++ tests/tests/compiletest.rs
@@ -5,7 +5,7 @@ use std::{env, path::PathBuf};
 
 use compiletest_rs::{common::Mode, Config};
 use itertools::Itertools;
-use tests::{find_flux_path, rustc_flags, FLUX_FULL_COMPILATION, FLUX_SYSROOT};
+use tests::{default_rustc_flags, find_flux_path, FLUX_FULL_COMPILATION, FLUX_SYSROOT};
 
 fn config() -> Config {
     let bless = env::args().any(|arg| arg == "--bless");
@@ -18,20 +18,18 @@ fn config() -> Config {
 
 fn test_runner(_: &[&()]) {
     let mut config = config().tempdir();
+    let mut rustc_flags = default_rustc_flags();
+    rustc_flags.extend(["--emit=metadata".to_string()]);
 
-    config.target_rustcflags = Some(rustc_flags().join(" "));
+    config.target_rustcflags = Some(rustc_flags.join(" "));
 
     config.clean_rmeta();
     config.clean_rlib();
     config.strict_headers = true;
 
-    // Enable full compilation such that we generate artifacts when annotating tests with `@aux-build`
+    // Enable full compilation so that we generate artifacts when annotating tests with `@aux-build`
     env::set_var(FLUX_FULL_COMPILATION, "1");
     env::set_var(FLUX_SYSROOT, config.rustc_path.parent().unwrap());
-
-    // config.mode = Mode::Ui;
-    // config.src_base = PathBuf::from("tests/asdf");
-    // compiletest_rs::run_tests(&config);
 
     let path: PathBuf = ["tests", "pos"].iter().collect();
     if path.exists() {
