+++ tests/tests/neg/extern_specs/auxiliary/extern_spec_impl01_aux.rs
@@ -1,0 +1,18 @@
+pub trait MyTrait {
+    fn foo() -> i32;
+}
+
+pub trait OtherTrait {
+    fn foo() -> i32;
+}
+
+impl<T> MyTrait for Vec<T> {
+    fn foo() -> i32 {
+        10
+    }
+}
+impl<T> OtherTrait for Vec<T> {
+    fn foo() -> i32 {
+        10
+    }
+}
