+++ tests/tests/neg/extern_specs/extern_spec_impl00.rs
