+++ tests/tests/neg/extern_specs/extern_spec_iter00.rs
