+++ tests/tests/neg/extern_specs/extern_spec_struct.rs
