+++ tests/tests/neg/extern_specs/extern_spec_trait00.rs
