+++ tests/tests/pos/extern_specs/auxiliary/extern_spec_impl01_aux.rs
@@ -1,5 +1,3 @@
-use flux_rs::extern_spec;
-
 pub trait MyTrait {
     fn foo() -> i32;
 }
@@ -18,16 +16,4 @@ impl<T> OtherTrait for Vec<T> {
     fn foo() -> i32 {
         10
     }
-}
-
-// "extern" impl
-#[extern_spec]
-impl<T> MyTrait for Vec<T> {
-    #[flux::sig(fn() -> i32[10])]
-    fn foo() -> i32;
-}
-
-#[flux::sig(fn () -> i32[10])]
-pub fn test_ok() -> i32 {
-    <Vec<i32> as MyTrait>::foo()
 }
