+++ tests/tests/pos/extern_specs/extern_spec_impl00.rs
