+++ tests/tests/pos/extern_specs/extern_spec_impl03.rs
