+++ tests/tests/pos/extern_specs/extern_spec_macro.rs
