+++ tests/tests/pos/extern_specs/extern_spec_struct00.rs
