+++ tests/tests/pos/extern_specs/extern_spec_struct03.rs
