+++ tests/tests/pos/extern_specs/extern_spec_trait00.rs
